+++ gui/src/app/pages/NwbPage/viewPlugins/Units/UnitsItemView.tsx
@@ -1,5 +1,6 @@
-import { FunctionComponent } from "react";
+import { FunctionComponent, useMemo } from "react";
 import DynamicTableView from "../DynamicTable/DynamicTableView";
+import { useSelectedUnitIds } from "app/package/context-unit-selection";
 
 type Props = {
   width: number;
@@ -14,12 +15,23 @@ const UnitsItemView: FunctionComponent<Props> = ({
   path,
   condensed,
 }) => {
+  const { selectedUnitIds, unitIdSelectionDispatch } = useSelectedUnitIds();
+  const selectedUnitIdsList = useMemo(() => {
+    return [...selectedUnitIds];
+  }, [selectedUnitIds]);
   return (
     <DynamicTableView
       width={width}
       height={height}
       path={path}
       referenceColumnName={"id"}
+      selectedRowIds={selectedUnitIdsList}
+      setSelectedRowIds={(ids) => {
+        unitIdSelectionDispatch({
+          type: "SET_SELECTION",
+          incomingSelectedUnitIds: ids,
+        });
+      }}
     />
   );
 };